Find mismatches in dataset using Python
In this notebook, I use a fictional dataset to demonstrate how to find the mismatches in the implementation dataset in comparion to the prescribed matching rules from a second dataset. This is also a problem I run into very often in Data Analytics. In the notebook, there are two datasets: the one is a large dataset in which two columns denote the implementation of the modelling requirements; the other is the requirement. The goal is to test whether the implementation in the two columns strictly follows the requirements prescribed.